apps/voice-agent/requirements.txt
livekit-agents[openai,deepgram,cartesia,turn-detector]~=1.0
python-dotenv>=1.0
httpx>=0.27.0
pydantic>=2.7
